analytics/requirements.txt
pyspark
python-dotenv
pandas
